requirements.txt
alembic==1.3.2
Babel==2.7.0
blinker==1.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
MyApplication==0.1.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.12
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1

# requirements for Heroku
psycopg2==2.7.3.1
gunicorn==19.7.1
